# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parking Lot LObby",
    "description": "PLL Api Documentation",
    "termsOfService": "",
    "contact": {
      "name": "matrix marketer",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1"
  },
...
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
    },
    "schemas": {
...
      "AddQuickResponse":{
        "type":"object",
        "properties": {
          "message": {
            "type": "string",
            "required":true
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateQuickResponse":{
        "type":"object",
        "properties": {
          "message": {
            "type": "string",
            "required":true
          },
          "responseId":{
            "type": "string",
            "required":true
          }
        },
        "required": [
          "message",
          "responseId"
        ]
      },
      "WaitingToCheckIn":{
        "type":"object",
        "properties": {
          "patientId": {
            "type": "string",
            "required":true
          }
        },
        "required": [
          "patientId"
        ]
      },
      "CheckInToCheckOut":{
        "type":"object",
        "properties": {
          "patientId": {
            "type": "string",
            "required":true
          }
        },
        "required": [
          "patientId"
        ]
      },
      "NotifyPatient":{
        "type":"object",
        "properties": {
          "patientId": {
            "type": "string",
            "required":true
          },
          "notifyTime": {
            "type": "number",
            "required":true
          }
        },
        "required": [
          "patientId",
          "notifyTime"
        ]
      },
...
      "createProfile":{
        "type":"object",
        "required": ["first_name","last_name","fullNumber", "dob"],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "fullNumber": {
            "type": "string"
          },
          "dob": {
            "type": "string"
          },
          "visitNotes": {
            "type": "string"
          },
          "parkingSpot": {
            "type": "string"
          }
        }        
      }
    },
...
  },
  "paths": {
...
    "/common/fetchPatientList": {
      "get": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "fetch Patient List",
        "description": "This for fetch list for a clinic",
        "tags": [
          "Common List"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "This field for type: 1=> Waiting list, 2=> Check-In list, 3=> Client serve today list"
          }, 
          {
            "in": "query",
            "name": "searchBy",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "This field for type: 1=> Waiting list, 2=> Check-In list, 3=> Client serve today list"
          }
        ],
        "responses": {
          "$ref": "#/components/responses"
        }
      }
    },
    "/common/getQuickResponse": {
      "get": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "fetch clinic quick response messages List",
        "description": "This for fetch list for a clinic",
        "tags": [
          "Common List"
        ],
        "parameters": [
        ],
        "responses": {
          "$ref": "#/components/responses"
        }
      }
    },
    "/common/addQuickResponse": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "addQuickResponse",
        "description": "This endpoint for add quick Responses for clinic(provider)",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/AddQuickResponse"
              }
            }
          }
        }
      }
    },
    "/common/updateQuickResponse": {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update QuickResponse",
        "description": "This endpoint for update quick Responses for clinic(provider)",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/UpdateQuickResponse"
              }
            }
          }
        }
      }
    },
    "/common/fetchPatientChat": {
      "get": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "fetch patient messages List",
        "description": "This for fetch patient message.",
        "tags": [
          "Common List"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "patientId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "patient Id"
          }
        ],
        "responses": {
          "$ref": "#/components/responses"
        }
      }
    },
    "/common/waitingToCheckIn": {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "waiting list to check in List",
        "description": "Move patient from waiting list to check in List",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/WaitingToCheckIn"
              }
            }
          }
        }
      }
    }, 
    "/common/checkInToCheckOut": {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "check in to check out",
        "description": "Move patient from check-In  list to check-Out List",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/CheckInToCheckOut"
              }
            }
          }
        }
      }
    },
    "/common/notifyPatient": {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Notify Patient about check In time",
        "description": "start a countdown for pateint",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/NotifyPatient"
              }
            }
          }
        }
      }
    },
...
    "/common/addClient" : {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Add profile",
        "description": "This for add a new  patient",
        "tags": [
          "Common List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/createProfile"
              }
            }
          }
        }
      }
    }
  }
}